+++ db/flush.go
@@ -4,12 +4,15 @@ import (
 	"context"
 	"fmt"
 	"reflect"
+	"time"
 
 	sink "github.com/streamingfast/substreams-sink"
 	"go.uber.org/zap"
 )
 
 func (l *Loader) Flush(ctx context.Context, moduleHash string, cursor *sink.Cursor) (err error) {
+	startAt := time.Now()
+
 	tx, err := l.DB.BeginTx(ctx, nil)
 	if err != nil {
 		return fmt.Errorf("failed to being db transaction: %w", err)
@@ -22,8 +25,12 @@ func (l *Loader) Flush(ctx context.Context, moduleHash string, cursor *sink.Curs
 		}
 	}()
 
+	var entryCount int
 	for tableName, entries := range l.entries {
-		l.logger.Info("flushing table entries", zap.String("table_name", tableName), zap.Int("entry_count", len(entries)))
+		if l.tracer.Enabled() {
+			l.logger.Debug("flushing table entries", zap.String("table_name", tableName), zap.Int("entry_count", len(entries)))
+		}
+
 		for _, entry := range entries {
 			query, err := entry.query(l.getType)
 			if err != nil {
@@ -35,22 +42,26 @@ func (l *Loader) Flush(ctx context.Context, moduleHash string, cursor *sink.Curs
 			}
 
 			if _, err := tx.ExecContext(ctx, query); err != nil {
-				return fmt.Errorf("executing query:\n`%s`\n err: %w", query, err)
+				return fmt.Errorf("executing query %q: %w", query, err)
 			}
 		}
+
+		entryCount += len(entries)
+
 	}
 
+	entryCount += 1
 	cursorQuery := l.UpdateCursorQuery(moduleHash, cursor)
 	if _, err := tx.ExecContext(ctx, cursorQuery); err != nil {
-		return fmt.Errorf("executing query:\n`%s`\n err: %w", cursorQuery, err)
+		return fmt.Errorf("executing query %q: %w", cursorQuery, err)
 	}
 
 	if err := tx.Commit(); err != nil {
 		return fmt.Errorf("failed to commit db transaction: %w", err)
 	}
-
 	l.reset()
 
+	l.logger.Debug("flushed table(s) entries", zap.Int("table_count", len(l.entries)), zap.Int("entry_count", entryCount), zap.Duration("took", time.Since(startAt)))
 	return nil
 }
 
@@ -58,7 +69,6 @@ func (l *Loader) reset() {
 	for tableName := range l.entries {
 		l.entries[tableName] = map[string]*Operation{}
 	}
-	l.EntriesCount = 0
 }
 
 func (l *Loader) getType(tableName string, columnName string) (reflect.Type, error) {